[CREATE] Config 모듈 구조화(개발용, 실제 운영용, 테스트용 설정 구별)
